requirements.txt
binance-futures-connector==4.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
mysql-connector-python==9.2.0
pycryptodome==3.22.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
